+++ src/main.py
@@ -1,5 +1,5 @@
 from fastapi import FastAPI
-from src.routes import user
+from src.routes import user, auth
 
 app = FastAPI()
 
@@ -8,3 +8,4 @@ def health_check():
     return {"status": "ok"}
 
 app.include_router(user.router)
+app.include_router(auth.router)