+++ src/routes/user.py
@@ -5,6 +5,7 @@ from typing import List
 from src.controllers.user import handle_create_user, handle_get_user_list, handle_update_user, handle_delete_user
 from src.database import get_db
 from src.schemas.user import UserCreate, UserResponse, UserUpdate
+from src.utils.security import get_current_user
 
 router = APIRouter(prefix="/users", tags=["Users"])
 
@@ -16,7 +17,7 @@ def create_user(
     return handle_create_user(user_data, db)
 
 @router.get("/", response_model=List[UserResponse], status_code=200)
-def get_users(db: Session = Depends(get_db)):
+def get_users(db: Session = Depends(get_db), user_data: dict = Depends(get_current_user) ):
     return handle_get_user_list(db)
 
 @router.put("/{user_id}", response_model=UserResponse, status_code=200)