+++ src/utils/security.py
@@ -1,4 +1,15 @@
+from jose import jwt, JWTError
 import bcrypt
+from dotenv import load_dotenv
+
+from datetime import datetime, timedelta, timezone
+import os
+
+from fastapi import Depends, HTTPException, status
+from fastapi.security import HTTPBearer, HTTPAuthorizationCredentials
+
+SECRET_KEY = os.getenv("SECRET_KEY")
+security = HTTPBearer()
 
 def hash_password(password: str) -> str:
     salt = bcrypt.gensalt()
@@ -7,3 +18,28 @@ def hash_password(password: str) -> str:
 
 def password_decode(password: str, hashed_password: str) -> bool:
     return bcrypt.checkpw(password.encode("utf-8"), hashed_password.encode("utf-8"))
+
+def create_acess_token(data: dict, expires_delta: timedelta = None):
+    to_encode = data.copy()
+    expire = datetime.now(timezone.utc) + (expires_delta or timedelta(minutes=15))
+    to_encode.update({"exp": expire})
+    return jwt.encode(to_encode, SECRET_KEY, algorithm="HS256")
+
+def get_current_user(credentials: HTTPAuthorizationCredentials = Depends(security)):
+    token = credentials.credentials  
+
+    credentials_exception = HTTPException(
+        status_code=status.HTTP_401_UNAUTHORIZED,
+        detail="Could not validate credentials",
+        headers={"WWW-Authenticate": "Bearer"}
+    )
+    
+    try:
+        payload = jwt.decode(token, SECRET_KEY, algorithms=["HS256"])
+        user_id: str = payload.get("sub")
+        role: str = payload.get("role")
+        if user_id is None or role is None:
+            raise credentials_exception
+        return {"user_id": int(user_id), "role": role}
+    except JWTError:
+        raise credentials_exception